backend/requirements.txt
Flask
flask-cors
Requests
python-decouple